telegram-bot/requirements.txt
# HERMES MDM Steward Telegram Bot — Python dependencies
# Tested on Python 3.11+ (Apple Silicon, macOS)

# Telegram bot framework — async, actively maintained
python-telegram-bot==21.9

# Anthropic Python SDK — Claude API
anthropic==0.49.0

# Environment variable management
python-dotenv==1.0.1
